#535 fix file not found exceptions to clean up log. does not affect gameplay either way and some json files are empty
