+++ src/main/java/com/lothrazar/cyclicmagic/block/BlockConveyor.java
@@ -7,12 +7,8 @@ import com.lothrazar.cyclicmagic.util.UtilChat;
 import com.lothrazar.cyclicmagic.util.UtilEntity;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
-import net.minecraft.block.properties.IProperty;
-import net.minecraft.block.properties.PropertyDirection;
-import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.SoundEvents;
 import net.minecraft.item.ItemStack;
@@ -30,7 +26,6 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 public class BlockConveyor extends BlockBaseFlat implements IHasRecipe {
   private static final int RECIPE_OUTPUT = 8;
   protected static final AxisAlignedBB AABB = new AxisAlignedBB(0.0D, 0.0D, 0.0D, 1D, 0.03125D, 1D);
-  private static final PropertyDirection PROPERTYFACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
   private final static float ANGLE = 1;
   private static final float powerCorrection = 0.02F;
   public static enum SpeedType {
@@ -45,10 +40,8 @@ public class BlockConveyor extends BlockBaseFlat implements IHasRecipe {
   public BlockConveyor(SpeedType t) {
     super(Material.IRON);//, MapColor.GRASS
     this.setSoundType(SoundType.SLIME);
-    this.setHardness(2.0F).setResistance(2.0F);
     type = t;
     sound = SoundEvents.BLOCK_ANVIL_BREAK;
-    this.setHardness(2.0F).setResistance(2.0F);
     switch (type) {
       case LARGE:
         this.power = 0.32F;
@@ -112,38 +105,6 @@ public class BlockConveyor extends BlockBaseFlat implements IHasRecipe {
         }
       }
     }
-  }
-  //below is all for facing
-  @Override
-  public IBlockState getStateFromMeta(int meta) {
-    EnumFacing facing = EnumFacing.getHorizontal(meta);
-    return this.getDefaultState().withProperty(PROPERTYFACING, facing);
-  }
-  public EnumFacing getFacingFromState(IBlockState state) {
-    return (EnumFacing) state.getValue(PROPERTYFACING);
-  }
-  @Override
-  public int getMetaFromState(IBlockState state) {
-    EnumFacing facing = (EnumFacing) state.getValue(PROPERTYFACING);
-    int facingbits = facing.getHorizontalIndex();
-    return facingbits;
-  }
-  @Override
-  public IBlockState getActualState(IBlockState state, IBlockAccess worldIn, BlockPos pos) {
-    return state;
-  }
-  @Override
-  protected BlockStateContainer createBlockState() {
-    return new BlockStateContainer(this, new IProperty[] { PROPERTYFACING });
-  }
-  @Override
-  public IBlockState getStateForPlacement(World worldIn, BlockPos pos, EnumFacing blockFaceClickedOn, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer) {
-    // find the quadrant the player is facing
-    EnumFacing enumfacing = (placer == null) ? EnumFacing.NORTH : EnumFacing.fromAngle(placer.rotationYaw);
-    return this.getDefaultState().withProperty(PROPERTYFACING, enumfacing);
-  }
-  public IBlockState onBlockPlaced(World worldIn, BlockPos pos, EnumFacing blockFaceClickedOn, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer) {
-    return this.getStateForPlacement(worldIn, pos, blockFaceClickedOn, hitX, hitY, hitZ, meta, placer);//110 support
   }
   @Override
   @SideOnly(Side.CLIENT)
