+++ src/main/java/com/lothrazar/cyclicmagic/block/BlockLaunch.java
@@ -7,20 +7,27 @@ import com.lothrazar.cyclicmagic.util.UtilChat;
 import com.lothrazar.cyclicmagic.util.UtilEntity;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.IProperty;
+import net.minecraft.block.properties.PropertyDirection;
+import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.IRecipe;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class BlockLaunch extends BlockBaseFlat implements IHasRecipe {
+public class BlockLaunch extends BlockBaseFlat implements IHasRecipe {  private static final PropertyDirection PROPERTYFACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
+
   private final static float ANGLE = 90;
   private final static int RECIPE_OUT = 6;
   public static enum LaunchType {
@@ -34,7 +41,6 @@ public class BlockLaunch extends BlockBaseFlat implements IHasRecipe {
     super(Material.IRON);//same as BlockSlime 
     //    this.blockMapColor = MapColor.GRASS;
     this.setSoundType(SoundType.SLIME);
-    this.setHardness(2.0F).setResistance(2.0F);
     sound = s;
     type = t;
     switch (type) {
