+++ src/main/resources/assets/cyclicmagic/blockstates/fire.json
@@ -1,0 +1,2 @@
+{
+}
