+++ src/main/resources/assets/cyclicmagic/blockstates/plate_push_fast.json
@@ -1,5 +1,6 @@
 {
     "variants": {
+        "normal": { "model": "cyclicmagic:plate_push_fast" },
         "facing=north": { "model": "cyclicmagic:plate_push_fast" },
         "facing=east":  { "model": "cyclicmagic:plate_push_fast" ,"y":90},
         "facing=south": { "model": "cyclicmagic:plate_push_fast" ,"y":180},
