+++ src/main/resources/assets/cyclicmagic/models/item/item_pipe_sort.json
@@ -1,0 +1,10 @@
+{
+    "parent": "cyclicmagic:block/item_pipe_sort",
+	"display": {
+        "thirdperson": {
+            "rotation": [ 10, -45, 170 ],
+            "translation": [ 0, 1.5, -2.75 ],
+            "scale": [ 0.375, 0.375, 0.375 ]
+        }
+    }
+}
