+++ src/main/resources/assets/cyclicmagic/models/item/milk.json
@@ -1,0 +1,5 @@
+{
+    "parent": "item/generated",
+    "textures": {
+    }
+}
