+++ src/main/resources/assets/cyclicmagic/models/item/plate_launch_med.json
@@ -1,0 +1,6 @@
+{
+    "parent": "item/handheld",
+    "textures": {
+        "layer0": "cyclicmagic:blocks/plate_launch_med"
+    }
+}
